+++ components/frontend_react/webapp/src/utils/api.ts
@@ -60,8 +60,11 @@ const jobsEndpoint = envOrFail(
 )
 
 export const fetchAllChatModels =
-  (token: string) => (): Promise<string[] | undefined> => {
-    const url = `${endpoint}/chat/chat_types`
+  (token: string, isMultimodal?: boolean) => (): Promise<string[] | undefined> => {
+    let url = `${endpoint}/chat/chat_types`
+    if (isMultimodal !== undefined) {
+      url += `?is_multimodal=${isMultimodal}`
+    }
     const headers = { Authorization: `Bearer ${token}` }
     return axios.get(url, { headers }).then(path(["data", "data"]))
   }
@@ -147,6 +150,13 @@ export const resumeChat =
     return axios.post(url, data, { headers }).then(path(["data", "data"]))
   }
 
+export const fetchEmbeddingTypes =
+  (token: string, are_multimodal: boolean) => (): Promise<string[] | undefined | null> => {
+    let url = `${endpoint}/llm/embedding_types?is_multimodal=${are_multimodal}`
+    const headers = { Authorization: `Bearer ${token}` }
+    return axios.get(url, { headers }).then(path(["data", "data"]))
+  }
+
 export const fetchQuery =
   (token: string, queryId: string | null) => (): Promise<Query | undefined | null> => {
     if (!queryId) return Promise.resolve(null)
@@ -198,6 +208,7 @@ export const createQueryEngine =
         "associated_engines": queryEngine.child_engines,
         "manifest_url": queryEngine.manifest_url,
         "chunk_size": queryEngine.chunk_size,
+        "is_multimodal": queryEngine.is_multimodal ? "True" : "False",
       }
     }
     return axios.post(url, data, { headers }).then(path(["data", "data"]))