+++ components/frontend_streamlit/src/components/query_engine_select.py
@@ -37,3 +37,6 @@ def query_engine_select():
           tuple(query_engines.keys()))
       query_engine_id = query_engines[qe_name]["id"]
       st.session_state.query_engine_id = query_engine_id
+      query_engine_is_multimodal = query_engines[qe_name]["params"]\
+        .get("is_multimodal", "False")
+      st.session_state.query_engine_is_multimodal = query_engine_is_multimodal